Break out API specific validation
